requirements.txt
beautifulsoup4~=4.12.2
requests~=2.31.0
Flask~=2.3.2
Flask-Login~=0.6.2
Flask-WTF~=1.1.1
Flask-Compress~=1.13
pymongo~=4.3.3
python-dotenv~=1.0.0
stundenplan24-wrapper@ git+https://github.com/Belissimo-T/stundenplan24-wrapper@95f7309c
Werkzeug~=2.3.6
discord-webhook~=1.3.0
brotlicffi~=1.1.0.0
pywebpush~=1.14.0
py-vapid~=1.9.0